Fixing double extraction from Qt widgets to use a float extraction instead

+++ Source/StatsGenerator/TableModels/SGBetaTableModel.cpp
@@ -235,13 +235,13 @@ bool SGBetaTableModel::setData(const QModelIndex & index, const QVariant & value
   switch(col)
   {
     case BinNumber:
-      m_BinNumbers[row] = value.toDouble(&ok);
+      m_BinNumbers[row] = value.toFloat(&ok);
       break;
     case Alpha:
-      m_Alpha[row] = value.toDouble(&ok);
+      m_Alpha[row] = value.toFloat(&ok);
       break;
     case Beta:
-      m_Beta[row] = value.toDouble(&ok);
+      m_Beta[row] = value.toFloat(&ok);
       break;
     case LineColor:
       m_Colors[row] = value.toString();
